:recycle:: fix Player and Food

+++ Agar.io/Assets/Scripts/Model/Entity.cs
@@ -5,7 +5,7 @@ namespace Agario.Model
     public abstract class Entity
     {
         public Vector2 Position { get; set; }
-        public float Radius { get; set; }
+        public int Radius { get; set; }
         public EntityType EntityType { get; set; }
         public bool IsDead { get; set; }
     }
